refactor(calendar): refactor file structure
