+++ src/framework/theme/components/calendar/_calendar.component.theme.scss
@@ -1,0 +1,11 @@
+/**
+ * @license
+ * Copyright Akveo. All Rights Reserved.
+ * Licensed under the MIT License. See License.txt in the project root for license information.
+ */
+
+@mixin nb-calendar-theme() {
+  nb-calendar {
+    color: nb-theme(color-fg);
+  }
+}
