+++ src/framework/theme/components/calendar/calendar.module.ts
@@ -8,12 +8,12 @@ import { ModuleWithProviders, NgModule } from '@angular/core';
 
 import { NbSharedModule } from '../shared/shared.module';
 import { NbCalendarComponent } from './calendar.component';
-import { NbCalendarMonthViewComponent } from './partials/calendar-month-view.component';
+import { NbCalendarMonthViewComponent } from './components/calendar-month-view/calendar-month-view.component';
 import { NbCalendarModelFactoryService } from './models/calendar-model-factory.service';
 import { NbNativeDateTimeUtilService } from './service/native-date-time-util.service';
 import { NbDateTimeUtil } from './service/date-time-util.interface';
 import { NbCheckboxModule } from '../checkbox/checkbox.module';
-import { NbCalendarCellViewComponent } from './partials/calendar-cell-view.component';
+import { NbCalendarCellViewComponent } from './components/callendar-cell-view/calendar-cell-view.component';
 
 const NB_CALENDAR_PROVIDERS = [
   NbCalendarModelFactoryService,
