+++ src/framework/theme/components/calendar/components/calendar-month-view/calendar-month-view.component.scss
