+++ src/framework/theme/components/calendar/components/callendar-cell-view/calendar-cell-view.component.scss
