+++ src/framework/theme/styles/global/_components.scss
@@ -6,6 +6,7 @@
 
 @import '../../components/layout/layout.component.theme';
 @import '../../components/sidebar/sidebar.component.theme';
+@import '../../components/calendar/calendar.component.theme';
 @import '../../components/card/card.component.theme';
 @import '../../components/card/flip-card/flip-card.component.theme';
 @import '../../components/card/reveal-card/reveal-card.component.theme';
@@ -26,6 +27,7 @@
 
   @include nb-layout-theme();
   @include nb-sidebar-theme();
+  @include nb-calendar-theme();
   @include nb-card-theme();
   @include nd-reveal-card-theme();
   @include nd-flip-card-theme();
